requirements.txt
distro==1.9.0
markdown-it-py==3.0.0
mdurl==0.1.2
psutil==6.1.1
py-cpuinfo==9.0.0
Pygments==2.18.0
rich==13.9.4
screeninfo==0.8.1
